% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alba_boot}
\alias{alba_boot}
\title{Bootstrapped VBGF estimates for the \link[TropFishR]{alba} data set}
\format{
\subsection{\code{alba_boot}}{

A \code{lfqBoot} object with two levels:
\describe{
\item{\code{$bootRaw}}{A \code{data.frame} of fitted VBGF parameters
(columns) by resampling (rows).}
\item{\code{$seed}}{A \code{numeric} vector of seed values set prior to each
resampling call to \link[fishboot]{lfqResample}.}
}
}
}
\usage{
alba_boot
}
\description{
Bootstrapped VBGF estimates for the alba length frequency data
set as estimated by \code{\link[TropFishR]{ELEFAN_GA}}.
}
\keyword{datasets}
