% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alba_boot.R
\docType{data}
\name{alba_boot}
\alias{alba_boot}
\title{Bootstrapped VBGF estimates for the alba lfq dataset}
\format{A list of class `lfqBoot`}
\usage{
data(alba_boot)
}
\description{
Bootstrapped VBGF estimates for the alba lfq dataset as estimated by
  \code{\link[TropFishR]{ELEFAN_GA}}.
}
\examples{
data(alba_boot)

head(alba_boot$bootRaw)
head(alba_boot$seed)

# plot
univariate_density(alba_boot)
LinfK_scatterhist(alba_boot, phi.contour = TRUE)


}
\keyword{bootstrap}
\keyword{data}
\keyword{dataset}
\keyword{length-frequency}
