% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFAN_SA_boot.R
\name{ELEFAN_SA_boot}
\alias{ELEFAN_SA_boot}
\title{Bootstraped ELEFAN_SA}
\usage{
ELEFAN_SA_boot(
  lfq,
  ...,
  nresamp = 200,
  resample = TRUE,
  seed = NULL,
  no_cores = 1,
  outfile = NA
)
}
\arguments{
\item{lfq}{a length frequency object of the class \code{lfq} (see
\link[TropFishR]{lfqCreate}).}

\item{...}{Extra arguments passed to \link[TropFishR]{ELEFAN_SA}:
\code{seasonalised}, \code{init_par}, \code{low_par}, \code{up_par},
\code{SA_time}, \code{maxit}, \code{nb.stop.improvement}, \code{SA_temp},
\code{verbose}, \code{MA}, \code{addl.sqrt}, \code{agemax} and
\code{flagging.out}. Default values remain the same and just \code{plot} and
\code{plot.score} will be set as \code{FALSE}.}

\item{nresamp}{\code{numeric}, the number of permutations to run (by default
\code{nresamp = 200}).}

\item{resample}{\code{logical}. Do you want that \code{lfq} object be
resampled (\code{TRUE} by default).}

\item{seed}{seed value for random number reproducibility.}

\item{no_cores}{positive integer. If \code{no_cores} > 1, a 'parallel' package
cluster with that many cores is created.}

\item{outfile}{\code{character}; path of the file which will register the
progress of the permutation completions. If it is set as \code{false},
\code{NA} or \code{NULL}, no file will be created.}
}
\value{
An object of class `lfqBoot` containing 2 levels:
\describe{
  \item{\code{$bootRaw}}{A \code{data.frame} of fitted VBGF parameters
  (columns) by resampling (rows).}
  \item{\code{$seed}}{A \code{numeric} vector of seed values set prior to each
  resampling call to \link[fishboot]{lfqResample}.}
}
}
\description{
This function performs a bootstrapped fitting of von Bertalanffy
growth function (VBGF) via the \code{\link[TropFishR]{ELEFAN_SA}} function.
Most of the arguments are simply passed to the function within many
permutations (resampling) of the original \code{lfq} data.
}
\details{
If \code{resample = FALSE}, a \strong{partial bootstrap} is performed,
reflecting solution variation due only to the search algorithm.
}
\examples{
\dontrun{
# load data
data("alba", package = "TropFishR")

# Define settings (for demo only)
MA        <- 7
low_par   <- list(Linf = 8, K = 0.1, t_anchor = 0, C = 0, ts = 0)
up_par    <- list(Linf = 15, K = 5, t_anchor = 1, C = 1, ts = 1)
SA_time   <- 3
SA_temp   <- 1e5
nresamp   <- 12


# Parallel version
res <- ELEFAN_GA_boot(lfq = alba, MA = MA, seasonalised = FALSE,
                      init_par = NULL, up_par = up_par, low_par = low_par,
                      SA_time = SA_time, SA_temp = SA_temp,
                      nresamp = nresamp, seed = 1,
                      parallel = TRUE, no_cores = parallel::detectCores() - 2)

res


# Non-parallel version
res <- ELEFAN_GA_boot(lfq = alba, MA = MA, seasonalised = FALSE,
                      init_par = NULL, up_par = up_par, low_par = low_par,
                      SA_time = SA_time, SA_temp = SA_temp,
                      nresamp = nresamp, seed = 1,
                      parallel = FALSE)

res

# Plot scatterhist of Linf and K
LinfK_scatterhist(res = res)
}
}
