% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELEFAN_SA_boot.R
\name{ELEFAN_SA_boot}
\alias{ELEFAN_SA_boot}
\title{Bootstraped ELEFAN_SA}
\usage{
ELEFAN_SA_boot(
  lfq,
  seasonalised = FALSE,
  init_par = list(Linf = 50, K = 0.5, t_anchor = 0.5, C = 0, ts = 0),
  low_par = NULL,
  up_par = NULL,
  SA_time = 60 * 1,
  maxit = NULL,
  nb.stop.improvement = NULL,
  SA_temp = 1e+05,
  MA = 5,
  addl.sqrt = FALSE,
  agemax = NULL,
  seed = NULL,
  nresamp = 200,
  resample = TRUE,
  no_cores = 1,
  outfile = NA
)
}
\arguments{
\item{lfq}{a length frequency object of the class \code{lfq} (see
\link[TropFishR]{lfqCreate}).}

\item{seasonalised}{logical; indicating if the seasonalised von Bertalanffy
growth function should be applied (default: FALSE).}

\item{init_par}{a list providing the Initial values for the components to be
optimized. When set to NULL the following default values are used:
 \itemize{
  \item \strong{Linf} length infinity in cm (default is the maximum
  length class in the data),
  \item \strong{K} curving coefficient (default: 0.5),
  \item \strong{t_anchor} time point anchoring growth curves in year-length
  coordinate system, corresponds to peak spawning month (range: 0 to 1,
  default: 0.5),
  \item \strong{C} amplitude of growth oscillation (range: 0 to 1, default: 0),
  \item \strong{ts} summer point (ts = WP - 0.5) (range: 0 to 1, default: 0);
}}

\item{low_par}{a list providing the lower bounds for components. When set to
NULL the following default values are used:
 \itemize{
  \item \strong{Linf} length infinity in cm (default is calculated from maximum
  length class in the data),
  \item \strong{K} curving coefficient (default: 0.01),
  \item \strong{t_anchor} time point anchoring growth curves in year-length
  coordinate system, corresponds to peak spawning month (range: 0 to 1,
  default: 0),
  \item \strong{C} amplitude of growth oscillation (range: 0 to 1, default: 0),
  \item \strong{ts} summer point (ts = WP - 0.5) (range: 0 to 1, default: 0);
}}

\item{up_par}{a list providing the upper bounds for components. When set to
NULL the following default values are used:
 \itemize{
  \item \strong{Linf} length infinity in cm (default is calculated from maximum
  length class in the data),
  \item \strong{K} curving coefficient (default: 0.01),
  \item \strong{t_anchor} time point anchoring growth curves in year-length
  coordinate system, corresponds to peak spawning month (range: 0 to 1,
  default: 0),
  \item \strong{C} amplitude of growth oscillation (range: 0 to 1, default: 0),
  \item \strong{ts} summer point (ts = WP - 0.5) (range: 0 to 1, default: 0);
}}

\item{SA_time}{numeric; Maximum running time in seconds (default : 60 * 1).}

\item{maxit}{Integer. Maximum number of iterations of the algorithm. Default
is NULL.}

\item{nb.stop.improvement}{Integer. The program will stop when there is no
any improvement in 'nb.stop.improvement' steps. Default is NULL}

\item{SA_temp}{numeric; Initial value for temperature (default : 1e5).}

\item{MA}{number indicating over how many length classes the moving average
should be performed (defalut: 5, for more information see
\link[TropFishR]{lfqRestructure}).}

\item{addl.sqrt}{Passed to \link[TropFishR]{lfqRestructure}. Applied an
additional square-root transformation of positive values according to Brey
et al. (1988). (default: FALSE, for more information see
\link[TropFishR]{lfqRestructure}).}

\item{agemax}{maximum age of species; default NULL, then estimated from
\eqn{L_{inf}}.}

\item{seed}{seed value for random number reproducibility.}

\item{nresamp}{\code{numeric}, the number of permutations to run (by default
\code{nresamp = 200}).}

\item{resample}{\code{logical}. Do you want that \code{lfq} object be
resampled (\code{TRUE} by default).}

\item{no_cores}{positive integer. If \code{no_cores} > 1, a 'parallel' package
cluster with that many cores is created.}

\item{outfile}{\code{character}; path of the file which will register the
progress of the permutation completions. If it is set as \code{false},
\code{NA} or \code{NULL}, no file will be created.}
}
\value{
An object of class \code{lfqBoot} containing 2 levels:
\describe{
  \item{\code{$bootRaw}}{A \code{data.frame} of fitted VBGF parameters
  (columns) by resampling (rows).}
  \item{\code{$seed}}{A \code{numeric} vector of seed values set prior to each
  resampling call to \link[fishboot]{lfqResample}.}
}
}
\description{
This function performs a bootstrapped fitting of von Bertalanffy
growth function (VBGF) via the \code{\link[TropFishR]{ELEFAN_SA}} function.
Most of the arguments are simply passed to the function within many
permutations (resampling) of the original \code{lfq} data.
}
\details{
If \code{resample = FALSE}, a \strong{partial bootstrap} is performed,
reflecting solution variation due only to the search algorithm.
}
\examples{
\dontrun{
# load data
data("alba", package = "TropFishR")

# Define settings (for demo only)
MA        <- 7
low_par   <- list(Linf = 8, K = 0.1, t_anchor = 0, C = 0, ts = 0)
up_par    <- list(Linf = 15, K = 5, t_anchor = 1, C = 1, ts = 1)
SA_time   <- 3
SA_temp   <- 1e5
nresamp   <- 12


# Parallel version
res <- ELEFAN_GA_boot(lfq = alba, MA = MA, seasonalised = FALSE,
                      init_par = NULL, up_par = up_par, low_par = low_par,
                      SA_time = SA_time, SA_temp = SA_temp,
                      nresamp = nresamp, seed = NULL,
                      parallel = TRUE, no_cores = parallel::detectCores() - 2)

res


# Non-parallel version
res <- ELEFAN_GA_boot(lfq = alba, MA = MA, seasonalised = FALSE,
                      init_par = NULL, up_par = up_par, low_par = low_par,
                      SA_time = SA_time, SA_temp = SA_temp,
                      nresamp = nresamp, seed = NULL,
                      parallel = FALSE)

res

# Plot scatterhist of Linf and K
LinfK_scatterhist(res = res)
}
}
