% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishboot-package.R
\docType{package}
\name{fishboot-package}
\alias{fishboot}
\alias{fishboot-package}
\title{Bootstrap-based methods for the study of fish stocks and aquatic populations}
\description{
The fishboot package contains a suite of new bootstrap-based models
and software tools for the study of fish stocks and aquatic populations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rschwamborn/fishboot}
  \item Report bugs at \url{https://github.com/rschwamborn/fishboot/issues}
}

}
\author{
\strong{Maintainer}: Ralf Schwamborn \email{ralf.schwamborn@ufpe.br} (\href{https://orcid.org/0000-0001-9150-8720}{ORCID})

Authors:
\itemize{
  \item Tobias K. Mildenberger \email{tobm@aqua.dtu.dk}
  \item Marc H. Taylor \email{marc.taylor@thuenen.de}
  \item Margit Wilhelm \email{mwilhelm@unam.na} (\href{https://orcid.org/0000-0001-9271-2109}{ORCID})
}

Other contributors:
\itemize{
  \item Wencheng Lau-Medrano \email{luis.laum@gmail.com} (\href{https://orcid.org/0000-0002-7979-9710}{ORCID}) [contributor]
}

}
\keyword{internal}
