% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grolenage_boot.R
\name{grolenage_boot}
\alias{grolenage_boot}
\title{Bootstrapped length-at-age analysis}
\usage{
grolenage_boot(
  param,
  method = "LSM",
  Linf_est = NA,
  Linf_init = 10,
  K_init = 0.1,
  t0_init = 0,
  seed = as.numeric(Sys.time()),
  nresamp = 200,
  no_cores = 1,
  cc.only = FALSE
)
}
\arguments{
\item{param}{a \code{list} (or \code{data.frame}) consisting of following
parameters (levels/columns):
\itemize{
  \item \strong{age}: age measurements,
  \item \strong{length}: corresponding lengths in cm.
}}

\item{method}{indicating which of following methods should be applied:
\code{"GullandHolt"}, \code{"FordWalford"}, \code{"Chapman"},
\code{"BertalanffyPlot"}, or \code{"LSM"} (default).}

\item{Linf_est}{BertalanffyPlot requires an estimate for \eqn{L_{inf}} to
derive \eqn{K} and \eqn{t_0} (for more information see Details).}

\item{Linf_init}{initital parameter of \eqn{L_{inf}} for non-linear sqaures
fitting (default 10).}

\item{K_init}{initital parameter of \eqn{K} for non-linear sqaures fitting
(default 0.1).}

\item{t0_init}{initital parameter of \eqn{L_0} for non-linear squares fitting
(default 0).}

\item{seed}{seed value for random number reproducibility.}

\item{nresamp}{\code{numeric}; the number of permutations to run (Default:
\code{nresamp = 200}).}

\item{no_cores}{positive integer. If \code{no_cores} > 1, a 'parallel' package
cluster with that many cores is created.}

\item{cc.only}{\code{logical} Return complete cases only:  Do you want the
output object to omit results with failed estimates (with NaNs)? Remember
that if you define \code{cc.only = TRUE} the number of elements returned may
be less than specified in \code{nresamp}.}
}
\value{
An object of class \code{lfqBoot} containing 2 levels:
\describe{
  \item{\code{$bootRaw}}{A \code{data.frame} of fitted VBGF parameters
  (columns) by resampling (rows).}
  \item{\code{$seed}}{A \code{numeric} vector of seed values set prior to each
  resampling call.}
}

\code{t_anchor} (or \code{t0} in \link[TropFishR]{growth_length_age}) will be
only available for Bertalanffy plot and LSM method, otherwise a vector of
zeroes will be returned instead.
}
\description{
This function estimates growth parameters from length-at-age data.
Since it internally uses the function \link[TropFishR]{growth_length_age},
this function allows to perform different methods: Gulland and Holt, Ford
Walford, Chapman, Bertalanffy, or non linear least squares method (LSM).
}
\details{
CI and plotting arguments are set as \code{FALSE} for each bootstrap call
here.

By default, \link[TropFishR]{growth_length_age} generates a plot when is
called, so internally \code{grolenage_boot} executes a
\link[graphics]{dev.off} call in order to prevent it.

It is important to take into account the particular considerations of each
method regarding the required parameters, so it is recommended to read the
Details of the documentation of \link[TropFishR]{growth_length_age}.

Some words about \eqn{t_{anchor} = t_0}...
}
\examples{
# Synthetical length at age data
dat <- list(age = rep(x = 1:7,each = 5),
            length = c(rnorm(n = 5, mean = 25.7, sd = 0.9),
                       rnorm(n = 5, mean = 36.0, sd = 1.2),
                       rnorm(n = 5, mean = 42.9, sd = 1.5),
                       rnorm(n = 5, mean = 47.5, sd = 2.0),
                       rnorm(n = 5, mean = 50.7, sd = 0.4),
                       rnorm(n = 5, mean = 52.8, sd = 0.5),
                       rnorm(n = 5, mean = 54.2, sd = 0.7)))

growth_length_age(param = dat, method = "GullandHolt")

res <- grolenage_boot(param = dat, method = "GullandHolt")

# Plot scatterhist of Linf and K
LinfK_scatterhist(res = res)

# Bertalaffy plot
res <- grolenage_boot(dat, method = "BertalanffyPlot", Linf_est = 50,
                      nresamp = 100)

vbgfCI_time(res = res)
}
