% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grotag_boot.R
\name{grotag_boot}
\alias{grotag_boot}
\title{Resampling of growth increment data (\eqn{\frac{dL}{dt}}) from
mark-recapture (tagging) studies}
\usage{
grotag_boot(
  L1 = NULL,
  L2 = NULL,
  T1 = NULL,
  T2 = NULL,
  ...,
  input.data = NULL,
  nresamp = 200,
  seed = 123,
  no_cores = 1,
  cc.ony = FALSE
)
}
\arguments{
\item{L1, L2, T1, T2}{Name of the columns to be extracted from \code{input.data}
and used by the \link[fishmethods]{grotag} function for the arguments
\code{L1}, \code{L2}, \code{T1} and \code{T2}, respectively. See Details.}

\item{...}{Extra arguments passed to \link[fishmethods]{grotag}:
\code{alpha}, \code{beta}, \code{design}, \code{stvalue}, \code{upper},
\code{lower}, \code{st.ga}, \code{st.gb}, \code{st.galow}, \code{st.gaup},
\code{st.gblow}, \code{st.gbup} and \code{control}. It is important for the
user to indicate at least values of \code{alpha} and \code{beta}.}

\item{input.data}{A growth increment object of the class \code{data.frame}.}

\item{nresamp}{\code{numeric}; the number of permutations to run (Default:
\code{nresamp = 200}).}

\item{seed}{seed value for random number reproducibility.}

\item{no_cores}{positive integer. If \code{no_cores} > 1, a 'parallel' package
cluster with that many cores is created.}

\item{cc.ony}{\code{logical} Do you want the output object to omit results
with failed estimates (with NAs)? Remember that if you define
\code{cc.only = TRUE} the number of elements returned may be less than
specified in \code{nresamp}.}
}
\value{
A \code{data.frame} of fitted VBGF parameters (columns) by resampling
(rows). It includes a column (\code{seed}) with seed values set prior to each
resampling call.
}
\description{
This function resamples the \code{input.data} data by rows (i.e.,
by recapture date) several times (\code{nresamp} times, default:
\code{nresamp = 200}). Then, a VBGF curve is fitted to each resapled data set.
The output (a \code{list} of class \code{lfqBoot}) will store results (e.g.,
VGBGF function parameters K and Linf) in a \code{data.frame} accessible
through \code{$bootRaw}. The \code{$bootRaw} table also includes the growth
performance index \strong{Phi'}, seaonal parameters \strong{u} and \strong{w}
(sensu Francis, 1988), which are equal to \strong{C} (sensu Pauly and Gaschütz,
1979) and and \strong{ts} (sensu Mildenberger et al., 2017). The
\code{$bootRaw} table also includes seed values and system time.
}
\details{
There are 2 ways to specify the main input arguments (related to the size and
timing of the mark-recapture): (1) in the classical way, i.e. by defining
\code{L1}, \code{L2}, \code{T1} and \code{T2} as \code{numeric} vectors as
indicated in the \link[fishmethods]{grotag} documentation or (2) through a
\code{data.frame} indicated in the \code{input.data} argument. In the latter
case, the arguments \code{L1}, \code{L2}, \code{T1} and \code{T2} must be
1-length \code{character} vectors and they will serve to indicate the column
names of the corresponding variables. If only one value is specified for
\code{input.data} and any of the other arguments is NULL, a default name
equal to the variable name will be assigned (e.g. \code{L1 <- “L1”}).
}
\examples{
# Load example DB from fishmethods package
data(bonito, package = "fishmethods")

# Run the example cited on ?grotag
fishmethods::grotag(L1 = bonito$L1,
                    L2 = bonito$L2,
                    T1 = bonito$T1,
                    T2 = bonito$T2,
                    alpha = 35, beta = 55,
                    design  = list(nu = 1, m = 1,p = 1, sea = 1),
                    stvalue = list(sigma = 0.9, nu = 0.4, m = -1, p = 0.2, u = 0.4, w = 0.4),
                    upper   = list(sigma = 5, nu = 1, m = 2, p = 0.5, u = 1, w = 1),
                    lower   = list(sigma = 0, nu = 0, m = -2, p = 0.0, u = 0, w = 0),
                    control = list(maxit = 1e4))

# Run the example using grotag_boot
grotag_boot(L1 = bonito$L1,
            L2 = bonito$L2,
            T1 = bonito$T1,
            T2 = bonito$T2,
            alpha = 35, beta = 55,
            design  = list(nu = 1, m = 1,p = 1, sea = 1),
            stvalue = list(sigma = 0.9, nu = 0.4, m = -1, p = 0.2, u = 0.4, w = 0.4),
            upper   = list(sigma = 5, nu = 1, m = 2, p = 0.5, u = 1, w = 1),
            lower   = list(sigma = 0, nu = 0, m = -2, p = 0.0, u = 0, w = 0),
            control = list(maxit = 1e4),
            nresamp = 10)
}
