% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grotag_boot.R
\name{grotag_boot}
\alias{grotag_boot}
\title{Resampling of growth increment data (dL/dt) from mark-recapture (tagging) studies}
\usage{
grotag_boot(
  input.data,
  nresamp = 200,
  L1 = NULL,
  L2 = NULL,
  T1 = NULL,
  T2 = NULL,
  alpha = NULL,
  beta = NULL,
  design = list(nu = 0, m = 0, p = 0, sea = 0),
  stvalue = list(sigma = 0.9, nu = 0.4, m = -1, p = 0.01, u = 0.4, w = 0.4),
  upper = list(sigma = 5, nu = 1, m = 2, p = 1, u = 1, w = 1),
  lower = list(sigma = 0, nu = 0, m = -2, p = 0, u = 0, w = 0),
  gestimate = TRUE,
  st.ga = NULL,
  st.gb = NULL,
  st.galow = NULL,
  st.gaup = NULL,
  st.gblow = NULL,
  st.gbup = NULL,
  control = list(maxit = 10000)
)
}
\arguments{
\item{nresamp}{numeric; the number of permutations to run (Default: `nresamp = 200`)}

\item{L1}{Vector of length at release of tagged fish (see fishmethods::grotag)}

\item{L2}{Vector of length at recovery of tagged fish (see fishmethods::grotag)}

\item{T1}{Vector of julian time at release of tagged fish (see fishmethods::grotag)}

\item{T2}{Vector of julian time at recovery of tagged fish (see fishmethods::grotag)}

\item{alpha}{Numeric value giving an arbitrary length alpha (see fishmethods::grotag)}

\item{beta}{Numeric value giving an arbitrary length beta (beta> alpha) (see fishmethods::grotag)}

\item{design}{List specifying the design of the model to estimate. Use 1 to designate
ether a parameter(s) should be estimated. Type of parameters are: nu=growth
variability (1 parameter), m=bias parameter of measurement error (1 parameter),
p=outlier probability (1 parameter), and sea=seasonal variation (2 parameters: u and w). (see fishmethods::grotag)
Model 1 of Francis is the default settings of 0 for nu, m, p and sea. (see fishmethods::grotag)}

\item{stvalue}{Starting values of sigma (s) and depending on the design argument, nu, m, p, u, and w
used as input in the nonlinear estimation (function optim) routine.}

\item{upper}{Upper limit of the model parameters' (nu, m, p, u, and w) region to be
investigated.(see fishmethods::grotag)}

\item{lower}{Lower limit of the model parameters' (nu, m, p, u, and w) region to be
investigated. (see fishmethods::grotag)}

\item{gestimate}{Logical specifying whether starting values of ga and gb
(growth increments of alpha and beta) should be estimated automatically.
Default = TRUE.}

\item{st.ga}{If gestimate=FALSE, user-specified starting value for ga. (see fishmethods::grotag)}

\item{st.gb}{If gestimate=FALSE, user-specified starting value for gb. (see fishmethods::grotag)}

\item{st.galow}{If gestimate=FALSE, user-specified lower limit for st.ga used in optimization. (see fishmethods::grotag)}

\item{st.gaup}{If gestimate=FALSE, user-specified upper limit for st.ga used in optimization. (see fishmethods::grotag)}

\item{st.gblow}{If gestimate=FALSE, user-specified lower limit for st.gb used in optimization. (see fishmethods::grotag)}

\item{st.gbup}{If gestimate=FALSE, user-specified upper limit for st.gb used in optimization. (see fishmethods::grotag)}

\item{control}{Additional controls passed to the optimization function optim. (see fishmethods::grotag)}

\item{'increm'}{a growth increment object of the class `data.frame`.}
}
\value{
a list of class `lfqBoot` containing 2 levels:
`$bootRaw` - a data.frame of fitted VBGF parameters (columns) by resampling (rows),
`$seed` - a vector of seed values set prior to each resampling call to `Increm_Resample`.
}
\description{
The function resamples the `increm` data by rows
(i.e., by recapture date) several times ('nresamp' times, default: nresamp = 200).
Then, a VBGF curve is fitted to each resapled data set. The resulting
posteriors (e.g., VGBGF function parameters K and Linf) are stored in a data.frame
called '$bootRaw'.
The '$bootRaw' table also includes the growth performance index Phi',
seaonal parameters "u" and "w" sensu Francis,1988, whih are equal to "C"
sensu Pauly and Gaschütz, 1979 and and "ts" sensu Mildenberger et al., 2017.
The '$bootRaw' table also includes seed values and system time.
}
\examples{
library(fishmethods)
data(bonito)
res <- grotag_boot(bonito, nresamp = 10, L1=L1, L2=L2, T1=T1, T2=T2,alpha=35,beta=55, design=list(nu=1,m=1,p=1,sea=1), stvalue=list(sigma=0.9,nu=0.4,m=-1,p=0,u=0.4,w=0.4), upper=list(sigma=5,nu=1,m=2,p=0.5,u=1,w=1), lower=list(sigma=0,nu=0,m=-2,p=0,u=0,w=0),control=list(maxit=1e4))
res$bootRaw
res$seed
library(fishboot)
LinfK_scatterhist(res)
}
