% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinfK_scatterhist.R
\name{LinfK_scatterhist}
\alias{LinfK_scatterhist}
\title{Link/K scatterplot of bootstrapping results}
\usage{
LinfK_scatterhist(res, Linf.breaks = "Sturges", K.breaks = "Sturges",
  gridsize = rep(151, 2), H = ks::Hpi(res[, c("Linf", "K")]),
  shading = TRUE, shading.cols = colorRampPalette(c("white", blues9))(50),
  dens.contour = TRUE, probs = c(25, 50, 75, 95), phi.contour = FALSE,
  phi.levels = NULL, phi.contour.col = 8, phi.contour.lty = 2,
  phi.contour.lwd = 1, phi.contour.labcex = 0.75, pt.pch = 16,
  pt.col = rgb(0, 0, 0, 0.25), pt.cex = 0.5, pt.bg = 4,
  xlab = expression(italic("L")[infinity]), ylab = expression(italic("K")),
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{res}{list of class `lfqBoot`}

\item{Linf.breaks}{vector. Breaks for Linf histogram.}

\item{K.breaks}{vector. Breaks for K histogram.}

\item{gridsize}{vector. 2 values for defining the resolution of the grid}

\item{H}{object from `ks::H`}

\item{shading}{xx}

\item{shading.cols}{xx}

\item{dens.contour}{xx}

\item{probs}{xx}

\item{phi.contour}{xx}

\item{phi.levels}{xx}

\item{phi.contour.col}{xx}

\item{phi.contour.lty}{xx}

\item{phi.contour.lwd}{xx}

\item{phi.contour.labcex}{xx}

\item{pt.pch}{xx}

\item{pt.col}{xx}

\item{pt.cex}{xx}

\item{pt.bg}{xx}

\item{xlab}{xx}

\item{ylab}{xx}

\item{xlim}{xx}

\item{ylim}{xx}
}
\value{
plot
}
\description{
Link/K scatterplot of bootstrapping results
}
\examples{
data(alba_boot)
LinfK_scatterhist(alba_boot)

}
