% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbgfCI_time.R
\name{vbgfCI_time}
\alias{vbgfCI_time}
\title{VBGF plot and CI}
\usage{
vbgfCI_time(res, CI = 95, agemax = NULL, add_legend = TRUE,
  add_max_dens_legend = TRUE, n.eval.points = NULL,
  xlab = "Relative time", ylab = "Length",
  perm.col = adjustcolor("grey50", 0.1), perm.lwd = 1, ci.col = 1,
  ci.lty = 2, ci.lwd = 1, maxd.col = 1, maxd.lty = 1, maxd.lwd = 2)
}
\arguments{
\item{res}{list of class `lfqBoot`}

\item{CI}{xx}

\item{agemax}{xx}

\item{add_legend}{xx}

\item{add_max_dens_legend}{xx}

\item{n.eval.points}{xx}

\item{xlab}{xx}

\item{ylab}{xx}

\item{perm.col}{xx}

\item{perm.lwd}{xx}

\item{ci.col}{xx}

\item{ci.lty}{xx}

\item{ci.lwd}{xx}

\item{maxd.col}{xx}

\item{maxd.lty}{xx}

\item{maxd.lwd}{xx}
}
\value{
plot and data.frame with CI info
}
\description{
VBGF plot and CI
}
\examples{

data(alba_boot)
CIinfo <- vbgfCI_time(
  res = alba_boot,
  agemax = 2, CI = 95,
  perm.col = adjustcolor("grey50",0.2)
)


}
