% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbgfCI_time.R
\name{vbgfCI_time}
\alias{vbgfCI_time}
\title{VBGF plot and CI}
\usage{
vbgfCI_time(res, CI = 95, agemax = NULL, add_legend = TRUE,
  add_max_dens_legend = TRUE, xlab = "Relative time", ylab = "Length",
  perm.col = adjustcolor("grey50", 0.1), perm.lwd = 1, ci.col = 1,
  ci.lty = 2, ci.lwd = 1, maxd.col = 1, maxd.lty = 1, maxd.lwd = 2)
}
\arguments{
\item{res}{list of class `lfqBoot`}

\item{CI}{vector. Confidence interval levels to plot}

\item{agemax}{numeric. Maximum number of years to project.}

\item{add_legend}{logical. Should CI and max. density legend be added
(Default: `add_legend = TRUE`).}

\item{add_max_dens_legend}{logical. Should maximum density line be added
(Default: `add_max_dens_legend = TRUE`).}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{perm.col}{Color for each resample estimate line. See `?par`.}

\item{perm.lwd}{Line width for each resample estimate line. See `?par`.}

\item{ci.col}{Color for CI line. See `?par`.}

\item{ci.lty}{Line type for CI line. See `?par`.}

\item{ci.lwd}{Line width for CI line. See `?par`.}

\item{maxd.col}{Color for maximum density line. See `?par`.}

\item{maxd.lty}{Line type maximum density line. See `?par`.}

\item{maxd.lwd}{Line width maximum density line. See `?par`.}
}
\value{
plot and data.frame with CI info
}
\description{
VBGF plot and CI
}
\examples{

data(alba_boot)
CIinfo <- vbgfCI_time(
  res = alba_boot,
  agemax = 2, CI = 95,
  perm.col = adjustcolor("grey50",0.2)
)

# plot more CI levels
CIinfo <- vbgfCI_time(
  res = alba_boot,
  agemax = 2, CI = c(95, 50),
  ci.lty = 1, ci.lwd = 2, ci.col = c("red", "orange"),
  perm.col = adjustcolor("grey50",0.2)
)


}
